[refactor] useMarkerFeature 상태에 따른 분리
- 전역 상태를 상용하는 부분 / 그렇지 않은 부분에 대해 hook 분리
